Setup BSeq-SC External Dependencies

Configure bseqsc by setting up CIBERSORT source code.

bseqsc_config(file = NULL, error = FALSE)

Arguments

file

path to the CIBERSORT source R code.

error

logical that indicates if an error should be thrown if configuration failed.

Value

the path where the file was copied, or NULL if bseqsc is not correctly configured.

Details

BSeq-sc uses CIBERSORT to estimate cell type proportions, based on reference expression profiles. Due to licensing requirements, source code for this algorithm needs to be downloaded separately from its website http://cibersort.stanford.edu. It is released under the Stanford Non-Commercial License. In order to use it with bseqsc you will need to:

Got to http://cibersort.stanford.edu

Register and log in

Download the latest R source code from the http://cibersort.stanford.edu/download.php.

Configure bseqsc by pointing it to the downloaded file. This is done using the function bseqsc_config, which will copy the given R source file into the R-data/bseqsc sub-directory in the user's home directory for subsequent usage:

bseqsc_config('path/to/downloaded/source/CIBERSORT.R')