P-Value Scatter Plot

P-Value Scatter Plot

pvalueScatter(formula, data, pval.th = 0.05, pval.th.y = pval.th,
  pval.th.x = pval.th, label.th = Inf)

Arguments

formula

Formula of the form yvar ~ xvar that specify which variable in data contains variables for the x-axis and y-axis respectively.

data

a data.frame object containing the data to plot

pval.th

default p-value threshold used to filter te data shown in the plot. Only data with \(p <= pval.th\) are used.

pval.th.y, pval.th.x

p-value threshold to use for the variable in the x-axis and y-axis respectively

label.th

-log10(p-value) threshold above which data points are labelled. Only the points in the lower triangle are labelled.